+++ EssentialFeed/EssentialFeed/FeedImageDataStore.swift
@@ -1,0 +1,14 @@
+//
+//  FeedImageDataStore.swift
+//  EssentialFeed
+//
+//  Created by Prabhat Tiwari on 03/05/22.
+//
+
+import Foundation
+
+public protocol FeedImageDataStore {
+    typealias Result = Swift.Result<Data?, Error>
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (Result) -> Void)
+}
